backend/requirements.txt
annotated-types
anyio
bcrypt
certifi
cffi
charset-normalizer==3.4.2
click==8.2.1
cryptography==45.0.3
ecdsa==0.19.1
faiss-cpu==1.11.0
fastapi==0.115.12
filelock==3.18.0
fsspec==2025.5.1
h11==0.16.0
hf-xet==1.1.2
huggingface-hub==0.32.0
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
opencv-python==4.11.0.86
packaging==25.0
passlib==1.7.4
pillow==11.2.1
pyasn1==0.4.8
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
python-jose==3.4.0
python-multipart==0.0.20
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rsa==4.9.1
safetensors==0.5.3
setuptools==80.8.0
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
sympy==1.14.0
timm==1.0.15
tokenizers==0.21.1
torch==2.7.0
torchvision==0.22.0
tqdm==4.67.1
transformers==4.52.3
triton==3.3.0
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
